+++ executor/join.go
@@ -18,8 +18,10 @@ import (
 	"sync/atomic"
 
 	"github.com/juju/errors"
+	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/expression"
 	"github.com/pingcap/tidb/tablecodec"
+	"github.com/pingcap/tidb/terror"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/mvmap"
 	"github.com/pingcap/tidb/util/types"
@@ -409,4 +411,403 @@ func (e *HashJoinExec) Next() (Row, error) {
 	result := e.resultBuffer[e.resultCursor]
 	e.resultCursor++
 	return result, nil
+}
+
+// HashSemiJoinExec implements the hash join algorithm for semi join.
+type HashSemiJoinExec struct {
+	hashTable    map[string][]Row
+	smallHashKey []*expression.Column
+	bigHashKey   []*expression.Column
+	smallExec    Executor
+	bigExec      Executor
+	prepared     bool
+	ctx          context.Context
+	smallFilter  expression.CNFExprs
+	bigFilter    expression.CNFExprs
+	otherFilter  expression.CNFExprs
+	schema       *expression.Schema
+	resultRows   []Row
+	// auxMode is a mode that the result row always returns with an extra column which stores a boolean
+	// or NULL value to indicate if this row is matched.
+	auxMode           bool
+	smallTableHasNull bool
+	// anti is true, semi join only output the unmatched row.
+	anti bool
+}
+
+// Close implements the Executor Close interface.
+func (e *HashSemiJoinExec) Close() error {
+	e.hashTable = nil
+	e.resultRows = nil
+	return e.bigExec.Close()
+}
+
+// Open implements the Executor Open interface.
+func (e *HashSemiJoinExec) Open() error {
+	e.prepared = false
+	e.smallTableHasNull = false
+	e.hashTable = make(map[string][]Row)
+	e.resultRows = make([]Row, 1)
+	return errors.Trace(e.bigExec.Open())
+}
+
+// Schema implements the Executor Schema interface.
+func (e *HashSemiJoinExec) Schema() *expression.Schema {
+	return e.schema
+}
+
+// prepare runs the first time when 'Next' is called and it reads all data from the small table and stores
+// them in a hash table.
+func (e *HashSemiJoinExec) prepare() error {
+	err := e.smallExec.Open()
+	if err != nil {
+		return errors.Trace(err)
+	}
+	defer terror.Call(e.smallExec.Close)
+	e.hashTable = make(map[string][]Row)
+	e.resultRows = make([]Row, 1)
+	e.prepared = true
+	for {
+		row, err := e.smallExec.Next()
+		if err != nil {
+			return errors.Trace(err)
+		}
+		if row == nil {
+			return nil
+		}
+
+		matched, err := expression.EvalBool(e.smallFilter, row, e.ctx)
+		if err != nil {
+			return errors.Trace(err)
+		}
+		if !matched {
+			continue
+		}
+		hasNull, hashcode, err := getJoinKey(e.smallHashKey, row, make([]types.Datum, len(e.smallHashKey)), nil)
+		if err != nil {
+			return errors.Trace(err)
+		}
+		if hasNull {
+			e.smallTableHasNull = true
+			continue
+		}
+		if rows, ok := e.hashTable[string(hashcode)]; !ok {
+			e.hashTable[string(hashcode)] = []Row{row}
+		} else {
+			e.hashTable[string(hashcode)] = append(rows, row)
+		}
+	}
+}
+
+func (e *HashSemiJoinExec) rowIsMatched(bigRow Row) (matched bool, hasNull bool, err error) {
+	hasNull, hashcode, err := getJoinKey(e.bigHashKey, bigRow, make([]types.Datum, len(e.smallHashKey)), nil)
+	if err != nil {
+		return false, false, errors.Trace(err)
+	}
+	if hasNull {
+		return false, true, nil
+	}
+	rows, ok := e.hashTable[string(hashcode)]
+	if !ok {
+		return
+	}
+	// match eq condition
+	for _, smallRow := range rows {
+		matchedRow := makeJoinRow(bigRow, smallRow)
+		matched, err = expression.EvalBool(e.otherFilter, matchedRow, e.ctx)
+		if err != nil {
+			return false, false, errors.Trace(err)
+		}
+		if matched {
+			return
+		}
+	}
+	return
+}
+
+func (e *HashSemiJoinExec) fetchBigRow() (Row, bool, error) {
+	for {
+		bigRow, err := e.bigExec.Next()
+		if err != nil {
+			return nil, false, errors.Trace(err)
+		}
+		if bigRow == nil {
+			return nil, false, errors.Trace(e.bigExec.Close())
+		}
+
+		matched, err := expression.EvalBool(e.bigFilter, bigRow, e.ctx)
+		if err != nil {
+			return nil, false, errors.Trace(err)
+		}
+		if matched {
+			return bigRow, true, nil
+		} else if e.auxMode {
+			return bigRow, false, nil
+		}
+	}
+}
+
+func (e *HashSemiJoinExec) doJoin(bigRow Row, match bool) ([]Row, error) {
+	if e.auxMode && !match {
+		bigRow = append(bigRow, types.NewDatum(false))
+		e.resultRows[0] = bigRow
+		return e.resultRows, nil
+	}
+	matched, isNull, err := e.rowIsMatched(bigRow)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	if !matched && e.smallTableHasNull {
+		isNull = true
+	}
+	if e.anti && !isNull {
+		matched = !matched
+	}
+	// For the auxMode subquery, we return the row with a Datum indicating if it's a match,
+	// For the non-auxMode subquery, we return the matching row only.
+	if e.auxMode {
+		if isNull {
+			bigRow = append(bigRow, types.NewDatum(nil))
+		} else {
+			bigRow = append(bigRow, types.NewDatum(matched))
+		}
+		matched = true
+	}
+	if matched {
+		e.resultRows[0] = bigRow
+		return e.resultRows, nil
+	}
+	return nil, nil
+}
+
+// Next implements the Executor Next interface.
+func (e *HashSemiJoinExec) Next() (Row, error) {
+	if !e.prepared {
+		if err := e.prepare(); err != nil {
+			return nil, errors.Trace(err)
+		}
+	}
+
+	for {
+		bigRow, match, err := e.fetchBigRow()
+		if bigRow == nil || err != nil {
+			return bigRow, errors.Trace(err)
+		}
+		resultRows, err := e.doJoin(bigRow, match)
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		if len(resultRows) > 0 {
+			return resultRows[0], nil
+		}
+	}
+}
+
+// NestedLoopJoinExec implements nested-loop algorithm for join.
+type NestedLoopJoinExec struct {
+	innerRows     []Row
+	cursor        int
+	resultRows    []Row
+	SmallExec     Executor
+	BigExec       Executor
+	leftSmall     bool
+	prepared      bool
+	Ctx           context.Context
+	SmallFilter   expression.CNFExprs
+	BigFilter     expression.CNFExprs
+	OtherFilter   expression.CNFExprs
+	schema        *expression.Schema
+	outer         bool
+	defaultValues []types.Datum
+}
+
+// Schema implements Executor interface.
+func (e *NestedLoopJoinExec) Schema() *expression.Schema {
+	return e.schema
+}
+
+// Close implements Executor interface.
+func (e *NestedLoopJoinExec) Close() error {
+	e.resultRows = nil
+	e.innerRows = nil
+	return e.BigExec.Close()
+}
+
+// Open implements Executor Open interface.
+func (e *NestedLoopJoinExec) Open() error {
+	e.cursor = 0
+	e.prepared = false
+	e.resultRows = e.resultRows[:0]
+	e.innerRows = e.innerRows[:0]
+	return errors.Trace(e.BigExec.Open())
+}
+
+func (e *NestedLoopJoinExec) fetchBigRow() (Row, bool, error) {
+	for {
+		bigRow, err := e.BigExec.Next()
+		if err != nil {
+			return nil, false, errors.Trace(err)
+		}
+		if bigRow == nil {
+			return nil, false, e.BigExec.Close()
+		}
+
+		matched, err := expression.EvalBool(e.BigFilter, bigRow, e.Ctx)
+		if err != nil {
+			return nil, false, errors.Trace(err)
+		}
+		if matched {
+			return bigRow, true, nil
+		} else if e.outer {
+			return bigRow, false, nil
+		}
+	}
+}
+
+// prepare runs the first time when 'Next' is called and it reads all data from the small table and stores
+// them in a slice.
+func (e *NestedLoopJoinExec) prepare() error {
+	err := e.SmallExec.Open()
+	if err != nil {
+		return errors.Trace(err)
+	}
+	defer terror.Call(e.SmallExec.Close)
+	e.innerRows = e.innerRows[:0]
+	e.prepared = true
+	for {
+		row, err := e.SmallExec.Next()
+		if err != nil {
+			return errors.Trace(err)
+		}
+		if row == nil {
+			return nil
+		}
+
+		matched, err := expression.EvalBool(e.SmallFilter, row, e.Ctx)
+		if err != nil {
+			return errors.Trace(err)
+		}
+		if matched {
+			e.innerRows = append(e.innerRows, row)
+		}
+	}
+}
+
+func (e *NestedLoopJoinExec) fillRowWithDefaultValue(bigRow Row) (returnRow Row) {
+	smallRow := make([]types.Datum, e.SmallExec.Schema().Len())
+	copy(smallRow, e.defaultValues)
+	if e.leftSmall {
+		returnRow = makeJoinRow(smallRow, bigRow)
+	} else {
+		returnRow = makeJoinRow(bigRow, smallRow)
+	}
+	return returnRow
+}
+
+func (e *NestedLoopJoinExec) doJoin(bigRow Row, match bool) ([]Row, error) {
+	e.resultRows = e.resultRows[0:0]
+	if !match && e.outer {
+		row := e.fillRowWithDefaultValue(bigRow)
+		e.resultRows = append(e.resultRows, row)
+		return e.resultRows, nil
+	}
+	for _, row := range e.innerRows {
+		var mergedRow Row
+		if e.leftSmall {
+			mergedRow = makeJoinRow(row, bigRow)
+		} else {
+			mergedRow = makeJoinRow(bigRow, row)
+		}
+		matched, err := expression.EvalBool(e.OtherFilter, mergedRow, e.Ctx)
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		if !matched {
+			continue
+		}
+		e.resultRows = append(e.resultRows, mergedRow)
+	}
+	if len(e.resultRows) == 0 && e.outer {
+		e.resultRows = append(e.resultRows, e.fillRowWithDefaultValue(bigRow))
+	}
+	return e.resultRows, nil
+}
+
+// Next implements the Executor interface.
+func (e *NestedLoopJoinExec) Next() (Row, error) {
+	if !e.prepared {
+		if err := e.prepare(); err != nil {
+			return nil, errors.Trace(err)
+		}
+	}
+
+	for {
+		if e.cursor < len(e.resultRows) {
+			retRow := e.resultRows[e.cursor]
+			e.cursor++
+			return retRow, nil
+		}
+		bigRow, match, err := e.fetchBigRow()
+		if bigRow == nil || err != nil {
+			return bigRow, errors.Trace(err)
+		}
+		e.resultRows, err = e.doJoin(bigRow, match)
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		e.cursor = 0
+	}
+}
+
+// ApplyJoinExec is the new logic of apply.
+type ApplyJoinExec struct {
+	join        joinExec
+	outerSchema []*expression.CorrelatedColumn
+	cursor      int
+	resultRows  []Row
+	schema      *expression.Schema
+}
+
+// Schema implements the Executor interface.
+func (e *ApplyJoinExec) Schema() *expression.Schema {
+	return e.schema
+}
+
+// Close implements the Executor interface.
+func (e *ApplyJoinExec) Close() error {
+	return nil
+}
+
+// Open implements the Executor interface.
+func (e *ApplyJoinExec) Open() error {
+	e.cursor = 0
+	e.resultRows = nil
+	return errors.Trace(e.join.Open())
+}
+
+// Next implements the Executor interface.
+func (e *ApplyJoinExec) Next() (Row, error) {
+	for {
+		if e.cursor < len(e.resultRows) {
+			row := e.resultRows[e.cursor]
+			e.cursor++
+			return row, nil
+		}
+		bigRow, match, err := e.join.fetchBigRow()
+		if bigRow == nil || err != nil {
+			return nil, errors.Trace(err)
+		}
+		for _, col := range e.outerSchema {
+			*col.Data = bigRow[col.Index]
+		}
+		err = e.join.prepare()
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		e.resultRows, err = e.join.doJoin(bigRow, match)
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
+		e.cursor = 0
+	}
 }
